# solar-persistence-api/src/main/resources/solar-persistence-api-spec.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Solar persistence REST API",
    "description": "This API exposes APIs used by the Solar backend application and other Solar related applications.",
    "contact": {
      "name": "Fredrik Möller",
      "url": "https://github.com/afmoller/solar"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer-Authentication": []
    }
  ],
  "paths": {
...
    "/api/v1/weather/getByDate": {
      "get": {
        "tags": [
          "weather-controller"
        ],
        "operationId": "getWeatherByDateEntries",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherDataEntryDateDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "WeatherDataEntryDateDto": {
        "type": "object",
        "properties": {
          "dateUtc": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureIndoorInCelcius": {
            "type": "number",
            "format": "float"
          },
          "humidityIndoor": {
            "type": "integer",
            "format": "int32"
          },
          "baromAbsoluteIndoor": {
            "type": "number",
            "format": "float"
          },
          "temperatureOutdoorInCelsius": {
            "type": "number",
            "format": "float"
          },
          "humidityOutdoor": {
            "type": "integer",
            "format": "int32"
          },
          "windDirection": {
            "type": "integer",
            "format": "int32"
          },
          "windSpeedInMeterPerSecond": {
            "type": "number",
            "format": "float"
          },
          "windGustInMeterPerSecond": {
            "type": "number",
            "format": "float"
          },
          "maxDailyGustInMeterPerSecond": {
            "type": "number",
            "format": "float"
          },
          "uv": {
            "type": "integer",
            "format": "int32"
          },
          "solarRadiation": {
            "type": "number",
            "format": "float"
          },
          "rainRateInMillimetersPerHour": {
            "type": "number",
            "format": "float"
          },
          "hourlyRainInMillimeters": {
            "type": "number",
            "format": "float"
          },
          "dailyRainInMillimeters": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "dateUtc",
          "temperatureIndoorInCelcius",
          "humidityIndoor",
          "baromAbsoluteIndoor",
          "temperatureOutdoorInCelsius",
          "humidityOutdoor",
          "windDirection",
          "windSpeedInMeterPerSecond",
          "windGustInMeterPerSecond",
          "maxDailyGustInMeterPerSecond",
          "uv",
          "solarRadiation",
          "rainRateInMillimetersPerHour",
          "hourlyRainInMillimeters",
          "dailyRainInMillimeters"
        ]
      }
    },
    "securitySchemes": {
      "Bearer-Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}